requirements.txt
datasets>=2.18.0
langchain-community>=0.0.26
langchain-huggingface>=0.0.5
sentence-transformers>=2.6.1
qdrant-client>=1.8.0
numpy==1.26
PyYAML>=6.0
python-dotenv>=1.0.1
